Relocate nav items into own file
