+++ apps/airview-demo/src/layouts/main-layout/index.js
@@ -1,0 +1,1 @@
+export { MainLayout } from "./main-layout";
