+++ apps/airview-demo/src/layouts/main-layout/use-get-navigation-items-data.js
@@ -1,0 +1,5 @@
+import { mockNavItems } from "./mock-nav-items";
+
+export function useGetNavigationItemsData() {
+  return mockNavItems;
+}
